refactor: move common io-related functions into new file
